week02_management_and_testing/homework/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
backports.tarfile==1.2.0
build==1.2.2.post1
CacheControl==0.14.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
coverage==7.6.11
crashtest==0.4.1
cryptography==44.0.0
distlib==0.3.9
docker-pycreds==0.4.0
dulwich==0.22.7
exceptiongroup==1.2.2
fastjsonschema==2.21.1
filelock==3.17.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.8.0
Jinja2==3.1.5
keyring==25.6.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.61
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.2
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.12.0
platformdirs==4.3.6
pluggy==1.5.0
poetry==2.0.1
poetry-core==2.0.1
poetry-plugin-export==1.8.0
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyproject_hooks==1.2.0
pytest==7.4.4
pytest-cov==4.1.0
PyYAML==6.0.2
RapidFuzz==3.12.1
requests==2.32.3
requests-toolbelt==1.0.0
SecretStorage==3.3.3
sentry-sdk==2.20.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tomli==2.2.1
tomlkit==0.13.2
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
triton==2.1.0
trove-classifiers==2025.1.15.22
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.1
wandb==0.19.6
zipp==3.21.0
